pyproject.toml
[project]
    authors = [{ name = "Nate Landau", email = "github@natenate.org" }]
    dependencies = [
        "cappa>=0.26.6",
        "dynaconf>=3.2.10",
        "ffmpeg-progress-yield>=0.11.3",
        "ffmpeg-progress>=0.0.4",
        "ffmpeg-python>=0.2.0",
        "iso639-lang>=2.6.0",
        "requests>=2.32.3",
        "rich>=13.9.4",
        "shellingham>=1.5.4",
        "validators>=0.34.0",
    ]
    description = "Tools to transcode, inspect and convert videos."
    license = { file = "LICENSE" }
    name = "vid-cleaner"
    readme = "README.md"
    requires-python = ">=3.11"
    version = "0.3.6"

    [project.urls]
        Homepage   = "https://github.com/natelandau/vid-cleaner"
        Repository = "https://github.com/natelandau/vid-cleaner"

    [project.scripts]
        vidcleaner = "vid_cleaner.vidcleaner:main"

[build-system]
    build-backend = "hatchling.build"
    requires      = ["hatchling"]

[dependency-groups]
    dev = [
        "commitizen>=4.4.1",
        "coverage>=7.6.12",
        "duty>=1.6.0",
        "mypy>=1.15.0",
        "pre-commit>=4.1.0",
        "pytest-clarity>=1.0.1",
        "pytest-cov>=6.0.0",
        "pytest-env>=1.1.5",
        "pytest-mock>=3.14.0",
        "pytest-pretty-terminal>=1.1.0",
        "pytest-sugar>=1.0.0",
        "pytest-xdist>=3.6.1",
        "pytest>=8.3.5",
        "ruff>=0.9.9",
        "types-requests>=2.32.0.20250301",
        "typos>=1.30.1",
        "yamllint>=1.35.1",
    ]

[tool.commitizen]
    bump_message               = "bump(release): v$current_version → v$new_version"
    changelog_merge_prerelease = true
    tag_format                 = "v$version"
    update_changelog_on_bump   = true
    version_files              = ["src/vid_cleaner/constants.py:VERSION"]
    version_provider           = "pep621"

[tool.coverage.report] # https://coverage.readthedocs.io/en/latest/config.html#report
    exclude_lines = [
        'def __repr__',
        'except [\w\s\._]+ as .*:',
        'log\.critical',
        'log\.debug',
        'log\.error',
        'log\.exception',
        'log\.info',
        'log\.success',
        'log\.trace',
        'log\.warning',
        'pragma: no cover',
        'raise Abort',
        'raise Exit',
        'raise typer\.Exit',
    ]
    fail_under = 60
    precision = 1
    show_missing = true
    skip_covered = true

[tool.coverage.run]
    branch       = true
    command_line = "--module pytest"
    data_file    = ".reports/.coverage"
    source       = ["src"]

[tool.coverage.xml]
    output = ".reports/coverage.xml"

[tool.mypy] # https://mypy.readthedocs.io/en/latest/config_file.html
    disallow_any_unimported = false
    disallow_subclassing_any = false
    disallow_untyped_decorators = false
    disallow_untyped_defs = true
    exclude = [
        'tests/', # TOML literal string (single-quotes, regex okay, no escaping necessary)
    ]
    follow_imports = "normal"
    ignore_missing_imports = true
    junit_xml = ".reports/mypy.xml"
    no_implicit_optional = true
    pretty = false
    show_column_numbers = true
    show_error_codes = true
    show_error_context = true
    strict_optional = false
    warn_redundant_casts = true
    warn_unreachable = true
    warn_unused_ignores = true

[tool.pytest.ini_options]
    addopts        = "--color=yes --doctest-modules --exitfirst --failed-first --strict-config --strict-markers --junitxml=.reports/pytest.xml"
    filterwarnings = ["error", "ignore::DeprecationWarning"]
    testpaths      = ["src", "tests"]
    xfail_strict   = true

[tool.ruff] # https://github.com/charliermarsh/ruff
    exclude = [
        ".bzr",
        ".direnv",
        ".eggs",
        ".git",
        ".hg",
        ".mypy_cache",
        ".nox",
        ".pants.d",
        ".pytype",
        ".ruff_cache",
        ".svn",
        ".tox",
        ".venv",
        "__pypackages__",
        "_build",
        "buck-out",
        "build",
        "dist",
        "node_modules",
        "venv",
    ]

    fix            = true
    line-length    = 100
    output-format  = "grouped"
    src            = ["src", "tests"]
    target-version = "py310"

    [tool.ruff.lint]
        ignore = [
            "B006",   # mutable-argument-default
            "B008",   # function-call-in-default-argument
            "COM812", # Trailing comma missing"
            "CPY001", # Missing copyright notice at top of file
            "D107",   # undocumented-public-init
            "D213",   # multi-line-summary-second-line
            "D215",   # section-underline-not-over-indented
            "D406",   # new-line-after-section-name
            "D407",   # dashed-underline-after-section
            "D408",   # section-underline-after-name
            "D409",   # section-underline-matches-section-length
            "D413",   # blank-line-after-last-section
            "E501",   # line-too-long
            # "FBT001",  # Boolean default positional argument in function definition
            # "FBT002",  # Boolean default positional argument in function definition
            "ISC001",  # single-line-implicit-string-concatenation
            "N805",    # invalid-first-argument-name-for-method
            "PLR0913", # too-many-arguments
            "PLR0917", # too many positional arguments
            "RSE102",  # Unnecessary parentheses on raised exception"
            "S311",    # suspicious-non-cryptographic-random-usage
            "UP007",   # non-pep604-annotation
            # "UP045",   # Use `X | None` for type annotations
        ]
        per-file-ignores = { "vid_cleaner.py" = [
            "DOC201",
            "DOC501",
            "PLR0912",
            "PLR0913",
        ], "tests/*.py" = [
            "A002",
            "A003",
            "ANN001",
            "ANN003",
            "ANN201",
            "ANN202",
            "ARG001",
            "ARG002",
            "ARG005",
            "D102",
            "DOC201",
            "ERA001",
            "F403",
            "PGH003",
            "PLR0913",
            "PLR0917",
            "PLR2004",
            "PLR6301",
            "S101",
        ] }
        preview = true
        select = ["ALL"]
        unfixable = ["E712", "ERA001", "F401", "F841", "UP007"]
        [tool.ruff.lint.mccabe]
            # Unlike Flake8, default to a complexity level of 10.
            max-complexity = 10

        [tool.ruff.lint.pydocstyle]
            convention = "google"

        [tool.ruff.lint.pylint]
            max-args = 6

    [tool.ruff.format]
        indent-style              = "space"
        line-ending               = "auto"
        quote-style               = "double"
        skip-magic-trailing-comma = false
